+++ src/main/java/com/epam/homework4/cinemawebapp/service/IFilmService.java
@@ -5,13 +5,13 @@ import com.epam.homework4.cinemawebapp.model.Film;
 import java.util.List;
 
 public interface IFilmService {
-    Film getFilmById(int id);
+    Film getFilmById(Long id);
 
     List<Film> listFilms();
 
     Film createFilm(Film film);
 
-    Film updateFilm(int id, Film film);
+    Film updateFilm(Long id, Film film);
 
-    void deleteFilm(int id);
+    void deleteFilm(Long id);
 }
