+++ src/main/java/com/epam/homework4/cinemawebapp/service/IHallService.java
@@ -6,13 +6,13 @@ import com.epam.homework4.cinemawebapp.model.CinemaHall;
 import java.util.List;
 
 public interface IHallService {
-    CinemaHall getHallById(int id);
+    CinemaHall getHallById(Long id);
 
     List<CinemaHall> listHalls();
 
     CinemaHall createHall(CinemaHall hall);
 
-    CinemaHall updateHall(int id, CinemaHall hall);
+    CinemaHall updateHall(Long id, CinemaHall hall);
 
-    void deleteHall(int id);
+    void deleteHall(Long id);
 }
